warnings and code style, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12468 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/activitygen/city/AGCity.cpp
@@ -192,7 +192,7 @@ AGCity::generatePopulation() {
     for (it = streets.begin(); it != streets.end(); ++it) {
         people += it->getPopulation();
         while (people > 0 && idHouseholds < numAdults.size()) {
-            int i = RandHelper::rand(numAdults.size() - idHouseholds);
+            size_t i = RandHelper::rand(numAdults.size() - idHouseholds);
             ++idHouseholds;
             households.push_back(AGHousehold(&*it, this, idHouseholds));
             households.back().generatePeople(abs(numAdults[i]), numChilds[i], numAdults[i] < 0); //&statData
